refactor: More organizational shifts
This makes the top level program very closely match any components we
build, and avoids some "unused code" warnings that were cropping up.
