--- src/lib.rs
@@ -1,11 +1,0 @@
-mod block_inspector;
-mod message;
-pub mod settings;
-mod state;
-
-pub use message::*;
-pub use state::*;
-
-pub fn theme(_state: &State) -> iced::Theme {
-    iced::Theme::CatppuccinMocha
-}
