+++ src/main.rs
@@ -1,32 +1,33 @@
-use iced::{widget::column, window, Element, Subscription, Task};
-use pluton::Message;
+use iced::{window, Element, Subscription, Task};
 mod block_inspector;
+
+mod settings;
+mod state;
+mod update;
+mod view;
+
+use settings::*;
+use state::*;
+use update::*;
 
 fn main() -> iced::Result {
     iced::application("", update, view)
-        .theme(pluton::theme)
+        .theme(theme)
         .subscription(subscription)
-        .window(pluton::settings::window())
+        .window(window())
         .run()
 }
 
-fn subscription(_state: &pluton::State) -> Subscription<Message> {
+fn update(state: &mut State, message: Message) -> Task<Message> {
+    state.update(message)
+}
+
+fn view(state: &State) -> Element<Message> {
+    state.view()
+}
+fn theme(_state: &State) -> iced::Theme {
+    iced::Theme::CatppuccinMocha
+}
+fn subscription(_state: &State) -> Subscription<Message> {
     window::resize_events().map(|(_id, size)| Message::WindowResized(size))
-}
-
-fn update(state: &mut pluton::State, message: Message) -> Task<Message> {
-    use pluton::Message::*;
-
-    match message {
-        WindowResized(size) => {
-            state.window_size = size;
-
-            Task::none()
-        }
-        BlockInspector(m) => state.block_inspector.update(m).map(Message::BlockInspector),
-    }
-}
-
-fn view(state: &pluton::State) -> Element<pluton::Message> {
-    column![state.block_inspector.view().map(Message::BlockInspector)].into()
 }
