+++ src/state.rs
@@ -2,7 +2,6 @@ use crate::block_inspector;
 
 #[derive(Default)]
 pub struct State {
-    pub count: u64,
     pub window_size: iced::Size,
     pub block_inspector: block_inspector::State,
 }
